Support custom Metals server version again
When Metals was migrated to Scala 2.13, this library became responsible
for determining whether artifacts for Scala 2.12 or 2.13 should be
downloaded. This decision was made by comparing the user-provided
version with the version of Metals that first introduced 2.13. The
versions are compared with semver.

At the same time, this component supports passing complete Maven
coordinates (ie. `org:name:version`), which the semver library will not
be able to parse. Because parsing fails, Metals will not be started and
an error will be reported.

This commit fixes this by not trying to guess the Scala version of the
artifacts to download when complete coordinates are given.